@@ -2,10 +2,21 @@ import 'package:flutter/material.dart';
 import 'package:flutter_toast_pro/src/model/effect_type.dart';
 import 'model/message_type.dart';
 
-///显示消息的声明
-///[message] 消息的内容
-///[extra] 扩展数据
-///只用于普通消息
+/// Toast message builder.
+///
+/// message 类型 toast 的自定义构建器。
+///
+/// Parameters / 参数：
+/// - [context] BuildContext / 构建上下文
+/// - [message] Message text / 消息文本
+/// - [type] Message type / 消息类型
+/// - [effectType] Effect type / 效果类型
+/// - [alignment] Alignment in overlay / 覆盖层对齐
+/// - [extra] Extra data / 扩展数据
+///
+/// Only used for message toasts.
+///
+/// 仅用于 message 类型 toast。
 typedef ToastMessageBuilder =
     Widget Function(
       BuildContext context,
@@ -16,11 +27,20 @@ typedef ToastMessageBuilder =
       Map<String, dynamic> extra,
     );
 
-///显示进度消息的声明
-///[message] 加载中消息的内容
-///[progress] 进度数据
-///[extra] 扩展数据
-///只用于进度消息
+/// Toast progress builder.
+///
+/// progress 类型 toast 的自定义构建器。
+///
+/// Parameters / 参数：
+/// - [context] BuildContext / 构建上下文
+/// - [progress] Progress value / 进度值
+/// - [message] Optional message / 可选文案
+/// - [alignment] Alignment in overlay / 覆盖层对齐
+/// - [extra] Extra data / 扩展数据
+///
+/// Only used for progress toasts.
+///
+/// 仅用于 progress 类型 toast。
 typedef ToastProgressBuilder =
     Widget Function(
       BuildContext context,
@@ -30,10 +50,19 @@ typedef ToastProgressBuilder =
       Map<String, dynamic> extra,
     );
 
-///显示加载中的声明
-///[message] 消息的内容
-///[extra] 扩展数据
-///只用于加载中消息
+/// Toast loading builder.
+///
+/// loading 类型 toast 的自定义构建器。
+///
+/// Parameters / 参数：
+/// - [context] BuildContext / 构建上下文
+/// - [message] Optional message / 可选文案
+/// - [alignment] Alignment in overlay / 覆盖层对齐
+/// - [extra] Extra data / 扩展数据
+///
+/// Only used for loading toasts.
+///
+/// 仅用于 loading 类型 toast。
 typedef ToastLoadingBuilder =
     Widget Function(
       BuildContext context,
@@ -42,7 +71,12 @@ typedef ToastLoadingBuilder =
       Map<String, dynamic> extra,
     );
 
-///消息文本样式构建器声明
-///[type] 消息类型
-///[effectType] 效果类型
-typedef MessageTextStyleBuilder = TextStyle Function(MessageType type, EffectType effectType);
+/// Message text style builder.
+///
+/// message 文本样式构建器。
+///
+/// Parameters / 参数：
+/// - [type] Message type / 消息类型
+/// - [effectType] Effect type / 效果类型
+typedef MessageTextStyleBuilder =
+    TextStyle Function(MessageType type, EffectType effectType);
