@@ -1,22 +1,48 @@
 import 'package:flutter/material.dart';
+
 import '../flutter_toast_pro.dart';
 import 'model/toast_data_model.dart';
-import 'toast_event.dart';
 import 'model/toast_type.dart';
+import 'toast_event.dart';
 
+/// Public controller API for FlutterToastPro.
+///
+/// FlutterToastPro 的对外控制器 API。
 sealed class FlutterToastPro {
-  ///显示消息
+  /// Show a message toast.
+  ///
+  /// 显示 message toast。
   static void showMessage(
     String message, {
+
+    /// Override auto close duration for this message.
+    ///
+    /// 覆盖本次 message 的自动关闭时长。
     Duration? closeDuration,
+
+    /// Message type.
+    ///
+    /// 消息类型。
     MessageType type = MessageType.info,
+
+    /// Effect type override for this message.
+    ///
+    /// 覆盖本次 message 的效果类型。
     EffectType? effectType,
+
+    /// Alignment override for this toast.
+    ///
+    /// 覆盖本次 toast 的对齐方式。
     Alignment? alignment,
+
+    /// Extra data passed to custom builders.
+    ///
+    /// 传递给自定义 builder 的扩展数据。
     Map<String, dynamic> extra = const <String, dynamic>{},
   }) {
     ToastEvent.showMessages.add(
       ToastDataModel(
-        type: .message,
+        type: ToastType.message,
         message: message,
         messageType: type,
         effectType: effectType,
@@ -27,19 +53,37 @@ sealed class FlutterToastPro {
     );
   }
 
-  ///显示成功消息
+  /// Show a success message toast.
+  ///
+  /// 显示 success message toast。
   static void showSuccessMessage(
     String message, {
+
+    /// Override auto close duration for this message.
+    ///
+    /// 覆盖本次 message 的自动关闭时长。
     Duration? closeDuration,
+
+    /// Effect type override for this message.
+    ///
+    /// 覆盖本次 message 的效果类型。
     EffectType? effectType,
+
+    /// Alignment override for this toast.
+    ///
+    /// 覆盖本次 toast 的对齐方式。
     Alignment? alignment,
+
+    /// Extra data passed to custom builders.
+    ///
+    /// 传递给自定义 builder 的扩展数据。
     Map<String, dynamic> extra = const <String, dynamic>{},
   }) {
     ToastEvent.showMessages.add(
       ToastDataModel(
-        type: .message,
+        type: ToastType.message,
         message: message,
-        messageType: .success,
+        messageType: MessageType.success,
         effectType: effectType,
         closeDuration: closeDuration,
         alignment: alignment,
@@ -48,19 +92,41 @@ sealed class FlutterToastPro {
     );
   }
 
-  ///显示警告消息
+  /// Show a warning message toast.
+  ///
+  /// 显示 warning message toast。
+  ///
+  /// Note: API name kept as `showWaringMessage` for backward compatibility.
+  ///
+  /// 注意：为兼容历史 API，这里保留 `showWaringMessage` 的命名。
   static void showWaringMessage(
     String message, {
+
+    /// Override auto close duration for this message.
+    ///
+    /// 覆盖本次 message 的自动关闭时长。
     Duration? closeDuration,
+
+    /// Effect type override for this message.
+    ///
+    /// 覆盖本次 message 的效果类型。
     EffectType? effectType,
+
+    /// Alignment override for this toast.
+    ///
+    /// 覆盖本次 toast 的对齐方式。
     Alignment? alignment,
+
+    /// Extra data passed to custom builders.
+    ///
+    /// 传递给自定义 builder 的扩展数据。
     Map<String, dynamic> extra = const <String, dynamic>{},
   }) {
     ToastEvent.showMessages.add(
       ToastDataModel(
-        type: .message,
+        type: ToastType.message,
         message: message,
-        messageType: .warning,
+        messageType: MessageType.warning,
         effectType: effectType,
         closeDuration: closeDuration,
         alignment: alignment,
@@ -69,19 +135,37 @@ sealed class FlutterToastPro {
     );
   }
 
-  ///显示成功消息
+  /// Show an error message toast.
+  ///
+  /// 显示 error message toast。
   static void showErrorMessage(
     String message, {
+
+    /// Override auto close duration for this message.
+    ///
+    /// 覆盖本次 message 的自动关闭时长。
     Duration? closeDuration,
+
+    /// Effect type override for this message.
+    ///
+    /// 覆盖本次 message 的效果类型。
     EffectType? effectType,
+
+    /// Alignment override for this toast.
+    ///
+    /// 覆盖本次 toast 的对齐方式。
     Alignment? alignment,
+
+    /// Extra data passed to custom builders.
+    ///
+    /// 传递给自定义 builder 的扩展数据。
     Map<String, dynamic> extra = const <String, dynamic>{},
   }) {
     ToastEvent.showMessages.add(
       ToastDataModel(
-        type: .message,
+        type: ToastType.message,
         message: message,
-        messageType: .error,
+        messageType: MessageType.error,
         effectType: effectType,
         closeDuration: closeDuration,
         alignment: alignment,
@@ -90,22 +174,41 @@ sealed class FlutterToastPro {
     );
   }
 
-  ///手动隐藏消息
-  ///如果手动关闭消息时自动关闭时间还未结束也会立即关闭
+  /// Manually hide message toast.
+  ///
+  /// 手动隐藏 message toast。
+  ///
+  /// If auto-close timer is still running, the toast will be closed immediately.
+  ///
+  /// 如果自动关闭计时器仍在运行，toast 也会立即关闭。
   static void hideMessage() {
     ToastEvent.hideMessages.add(ToastType.message);
   }
 
-  ///显示进度
+  /// Show progress toast.
+  ///
+  /// 显示 progress toast。
   static void showProgress(
     double progress, {
+
+    /// Optional message.
+    ///
+    /// 可选文案。
     String? message,
+
+    /// Alignment override for this toast.
+    ///
+    /// 覆盖本次 toast 的对齐方式。
     Alignment? alignment,
+
+    /// Extra data passed to custom builders.
+    ///
+    /// 传递给自定义 builder 的扩展数据。
     Map<String, dynamic> extra = const <String, dynamic>{},
   }) {
     ToastEvent.showMessages.add(
       ToastDataModel(
-        type: .progress,
+        type: ToastType.progress,
         message: message,
         progress: progress,
         alignment: alignment,
@@ -114,60 +217,46 @@ sealed class FlutterToastPro {
     );
   }
 
-  ///关闭进度
+  /// Hide progress toast.
+  ///
+  /// 隐藏 progress toast。
   static void hideProgress() {
     ToastEvent.hideMessages.add(ToastType.progress);
   }
 
-  ///显示加载中
+  /// Show loading toast.
+  ///
+  /// 显示 loading toast。
   static void showLoading({
+    /// Optional message.
+    ///
+    /// 可选文案。
     String? message,
+
+    /// Alignment override for this toast.
+    ///
+    /// 覆盖本次 toast 的对齐方式。
     Alignment? alignment,
+
+    /// Extra data passed to custom builders.
+    ///
+    /// 传递给自定义 builder 的扩展数据。
     Map<String, dynamic> extra = const <String, dynamic>{},
   }) {
     ToastEvent.showMessages.add(
-      ToastDataModel(type: .loading, message: message, alignment: alignment, extra: extra),
+      ToastDataModel(
+        type: ToastType.loading,
+        message: message,
+        alignment: alignment,
+        extra: extra,
+      ),
     );
   }
 
-  ///关闭加载中
+  /// Hide loading toast.
+  ///
+  /// 隐藏 loading toast。
   static void hideLoading() {
     ToastEvent.hideMessages.add(ToastType.loading);
-  }
-
-  ///默认的 message 颜色
-  static MessageStyleOptions defaultMessageStyleOptions() {
-    return MessageStyleOptions(
-      info: InfoMessageColorOption(
-        primaryLight: ColorOption(
-          backgroundColor: Colors.black.withValues(alpha: 0.3),
-          foregroundColor: Colors.white,
-        ),
-        primary: ColorOption(backgroundColor: Colors.black, foregroundColor: Colors.white),
-      ),
-      success: SuccessMessageColorOption(
-        primaryLight: ColorOption(
-          backgroundColor: Color(0xFF67c23a).withValues(alpha: 0.3),
-          foregroundColor: Color(0xFF67c23a),
-        ),
-        primary: ColorOption(backgroundColor: Color(0xFF67c23a), foregroundColor: Colors.white),
-      ),
-      warning: WarningMessageColorOption(
-        primaryLight: ColorOption(
-          backgroundColor: Color(0xFFe6a23d).withValues(alpha: 0.3),
-          foregroundColor: Color(0xFFe6a23d),
-        ),
-        primary: ColorOption(backgroundColor: Color(0xFFe6a23d), foregroundColor: Colors.white),
-      ),
-      error: ErrorMessageColorOption(
-        primaryLight: ColorOption(
-          backgroundColor: Color(0xFFEF4444).withValues(alpha: 0.3),
-          foregroundColor: Color(0xFFEF4444),
-        ),
-        primary: ColorOption(backgroundColor: Color(0xFFEF4444), foregroundColor: Colors.white),
-      ),
-      padding: EdgeInsetsGeometry.symmetric(horizontal: 16, vertical: 12),
-      borderRadius: BorderRadiusGeometry.circular(10),
-    );
   }
 }
