@@ -4,7 +4,13 @@ import '../model/effect_type.dart';
 import '../model/message_style_options.dart';
 import '../model/message_type.dart';
 
+/// Default message widget used when no custom [ToastMessageBuilder] is provided.
+///
+/// 当未提供自定义 [ToastMessageBuilder] 时使用的默认 message 组件。
 class DefaultMessageWidget extends StatelessWidget {
+  /// Create a default message widget.
+  ///
+  /// 创建默认 message 组件。
   const DefaultMessageWidget({
     super.key,
     required this.message,
@@ -14,77 +20,89 @@ class DefaultMessageWidget extends StatelessWidget {
     required this.styleOptions,
   });
 
+  /// Message text.
+  ///
+  /// 消息文本内容。
   final String message;
 
+  /// Message type (info/success/warning/error).
+  ///
+  /// 消息类型（info/success/warning/error）。
   final MessageType type;
 
+  /// Effect type (primary/primaryLight).
+  ///
+  /// 效果类型（primary/primaryLight）。
   final EffectType effectType;
 
+  /// Alignment of the message within the overlay.
+  ///
+  /// message 在覆盖层中的对齐位置。
   final AlignmentGeometry alignment;
 
+  /// Style options for message widget.
+  ///
+  /// message 组件样式配置。
   final MessageStyleOptions styleOptions;
+
+  /// Resolve the color tone based on [type] and [effectType].
+  ///
+  /// 根据 [type] 与 [effectType] 计算最终使用的颜色配置。
+  ColorOption _toneFor(MessageType type, EffectType effectType) {
+    switch (type) {
+      case MessageType.info:
+        return effectType == EffectType.primaryLight
+            ? styleOptions.info.primaryLight
+            : styleOptions.info.primary;
+      case MessageType.success:
+        return effectType == EffectType.primaryLight
+            ? styleOptions.success.primaryLight
+            : styleOptions.success.primary;
+      case MessageType.warning:
+        return effectType == EffectType.primaryLight
+            ? styleOptions.warning.primaryLight
+            : styleOptions.warning.primary;
+      case MessageType.error:
+        return effectType == EffectType.primaryLight
+            ? styleOptions.error.primaryLight
+            : styleOptions.error.primary;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
-    var size = MediaQuery.sizeOf(context);
-    Color bgColor;
-    Color textColor;
-    switch (type) {
-      case MessageType.info:
-        switch (effectType) {
-          case EffectType.primaryLight:
-            bgColor = styleOptions.info.primaryLight.backgroundColor;
-            textColor = styleOptions.info.primaryLight.foregroundColor;
-          case EffectType.primary:
-            bgColor = styleOptions.info.primary.backgroundColor;
-            textColor = styleOptions.info.primary.foregroundColor;
-        }
+    final size = MediaQuery.sizeOf(context);
 
-      case MessageType.success:
-        switch (effectType) {
-          case EffectType.primaryLight:
-            bgColor = styleOptions.success.primaryLight.backgroundColor;
-            textColor = styleOptions.success.primaryLight.foregroundColor;
-          case EffectType.primary:
-            bgColor = styleOptions.success.primary.backgroundColor;
-            textColor = styleOptions.success.primary.foregroundColor;
-        }
+    final ColorOption tone = _toneFor(type, effectType);
 
-      case MessageType.warning:
-        switch (effectType) {
-          case EffectType.primaryLight:
-            bgColor = styleOptions.warning.primaryLight.backgroundColor;
-            textColor = styleOptions.warning.primaryLight.foregroundColor;
-          case EffectType.primary:
-            bgColor = styleOptions.warning.primary.backgroundColor;
-            textColor = styleOptions.warning.primary.foregroundColor;
-        }
+    // Base text style without enforced color.
+    //
+    // 基础文本样式（不强制颜色）。
+    final TextStyle baseStyle =
+        styleOptions.textStyleBuilder?.call(type, effectType) ??
+        const TextStyle();
 
-      case MessageType.error:
-        switch (effectType) {
-          case EffectType.primaryLight:
-            bgColor = styleOptions.error.primaryLight.backgroundColor;
-            textColor = styleOptions.error.primaryLight.foregroundColor;
-          case EffectType.primary:
-            bgColor = styleOptions.error.primary.backgroundColor;
-            textColor = styleOptions.error.primary.foregroundColor;
-        }
-    }
     return Container(
       width: size.width,
       height: size.height,
       alignment: alignment,
-      padding: styleOptions.margin,
       child: Container(
+        margin: styleOptions.margin,
         padding: styleOptions.padding,
-        decoration: BoxDecoration(color: bgColor, borderRadius: styleOptions.borderRadius),
+        decoration: BoxDecoration(
+          color: tone.backgroundColor,
+          borderRadius: styleOptions.borderRadius,
+        ),
         child: Text(
           message,
-          style:
-              styleOptions.textStyleBuilder
-                  ?.call(type, effectType)
-                  .merge(TextStyle(color: textColor)) ??
-              TextStyle(color: textColor),
+          // Merge in resolved foreground color.
+          //
+          // merge 最终前景色。
+          style: baseStyle.merge(TextStyle(color: tone.foregroundColor)),
+          // Allow wrapping by default (messages can be long).
+          //
+          // 默认允许换行（message 可能较长）。
+          softWrap: true,
         ),
       ),
     );
